% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\docType{data}
\name{Cacher}
\alias{Cacher}
\title{Cache Data}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
cache = Cacher$new(expire=3600)

cache$set(key, value, expire)

cache$get(key)

cache$exist(key)

cache$get_callback(key, callback, ...)
}
\arguments{
\item{expire}{Cache time(seconds). The default value is 3600, which means one hour.}

\item{value}{value}

\item{callback}{A callback function used to set the cached value, if the given key is not existed.}
}
\description{
Offer a `Cacher` object to save data in a range of time.
}
\examples{
cache = Cacher$new(expire = 3) # cache 3 seconds
cache$set('a', 1)
cache$get('a') # 1
Sys.sleep(3)
cache$get('a') # NULL

# set expire time explicitly
cache$set('b', 1, expire=10)

# exist or not
cache$exist('b')  # return FALSE if it is expired

# `get_callback` if key exists, return the cached value, otherwise
# run the callback function and cache its returned value
cache$get_callback('b', function() 123, expire=60)
Sys.sleep(10)
cache$get_callback('b', function() 123, expire=60)

}
\keyword{datasets}
